+++ src/main/java/sn/ousoka/GestionFile/repository/ServiceRepository.java
@@ -4,5 +4,5 @@ import java.util.List;
 import sn.ousoka.GestionFile.model.OKService;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceRepository extends JpaRepository<OKService, Integer> {
+public interface ServiceRepository extends JpaRepository<OKService, Long> {
 }
